+++ pubrec.go
@@ -102,8 +102,8 @@ func (p *PubRec) UnmarshalBinary(data []byte) error {
 	return b.err
 }
 
-func (p *PubRec) propertyMap() map[Ident]wireType {
-	return map[Ident]wireType{
-		ReasonString: &p.reason,
+func (p *PubRec) propertyMap() map[Ident]func() wireType {
+	return map[Ident]func() wireType{
+		ReasonString: func() wireType { return &p.reason },
 	}
 }