+++ subscribe_test.go
@@ -30,7 +30,6 @@ func ExampleSubscribe_Dump() {
 	Dump(os.Stdout, s)
 	// output:
 	// PacketID: 0
-	// SubscriptionID: 0
 	// Filters
 	//   0. a/b/c --r0pn2-
 	//   1. d/e --r0---1
@@ -39,8 +38,9 @@ func ExampleSubscribe_Dump() {
 }
 
 func ExampleSubscribe_malformed() {
-	// no filters
-	fmt.Println(NewSubscribe())
+	{// no filters
+		fmt.Println(NewSubscribe())
+	}
 	{ // bad qos
 		s := NewSubscribe()
 		s.SetSubscriptionID(1)
@@ -53,12 +53,6 @@ func ExampleSubscribe_malformed() {
 		s.AddFilters(NewTopicFilter("", OptQoS1))
 		fmt.Println(s)
 	}
-	{ // missing subscription id
-		s := NewSubscribe()
-		s.SetSubscriptionID(0)
-		s.AddFilters(NewTopicFilter("#", OptQoS1))
-		fmt.Println(s)
-	}
 	{ // too large subscription id
 		s := NewSubscribe()
 		s.SetSubscriptionID(268_435_455 + 1)
@@ -69,7 +63,6 @@ func ExampleSubscribe_malformed() {
 	// SUBSCRIBE --1- p0  5 bytes, malformed! no filters
 	// SUBSCRIBE --1- p0 a/b --r0--!! 13 bytes, malformed! invalid QoS
 	// SUBSCRIBE --1- p0  --r0---1 10 bytes, malformed! empty filter
-	// SUBSCRIBE --1- p0 # --r0---1 9 bytes, malformed! missing sub ID
 	// SUBSCRIBE --1- p0 # --r0---1 15 bytes, malformed! too large sub ID
 }