+++ unsuback.go
@@ -81,7 +81,7 @@ func (p *UnsubAck) variableHeader(b []byte, i int) int {
 func (p *UnsubAck) properties(b []byte, i int) int {
 	n := i
 	for id, v := range p.propertyMap() {
-		i += v.fillProp(b, i, id)
+		i += v().fillProp(b, i, id)
 	}
 	i += p.UserProperties.properties(b, i)
 	return i - n
@@ -110,8 +110,8 @@ func (p *UnsubAck) UnmarshalBinary(data []byte) error {
 	return b.err
 }
 
-func (p *UnsubAck) propertyMap() map[Ident]wireType {
-	return map[Ident]wireType{
-		ReasonString: &p.reasonString,
+func (p *UnsubAck) propertyMap() map[Ident]func() wireType {
+	return map[Ident]func() wireType{
+		ReasonString: func() wireType { return &p.reasonString },
 	}
 }