+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- Fix Subscribe.subscriptionID is optional
 - Add func Unsubscribe.Filters
 - Update dependencies